notifications, spotbilling, search etc

+++ modules/SpotBilling/model/domain-objects/MeterReading.js
@@ -41,7 +41,7 @@ class MeterReading extends DomainObject {
         return {
             meter_no: '*',
             current_reading: 'numeric',
-            'current_bill?': 'int'
+            'current_bill?': 'numeric'
         };
     }
